requirements.txt
awesome-slugify==1.6.5
backports.csv==1.0.6
certifi==2018.11.29
chardet==3.0.4
defusedxml==0.5.0
diff-match-patch==20181111
Django==1.9
django-braces==1.13.0
django-crispy-forms==1.7.2
django-formtools==2.1
django-haystack==2.8.1
django-import-export==1.2.0
django-polls==0.1
django-pure-pagination==0.3.0
django-simple-captcha==0.4.6
et-xmlfile==1.0.1
future==0.17.1
httplib2==0.9.2
idna==2.8
jdcal==1.4
jieba==0.39
Markdown==3.1
markdown2==2.3.7
MySQL-python==1.2.5
odfpy==1.4.0
openpyxl==2.6.1
Pillow==5.4.1
Pygments==2.3.1
pytz==2019.1
PyYAML==3.13
regex==2019.6.8
requests==2.21.0
six==1.12.0
tablib==0.13.0
Unidecode==0.4.21
urllib3==1.24.1
Whoosh==2.7.4
xlrd==1.2.0
xlwt==1.3.0
